FindCommand class: Add validation for keyword input
The current FindCommand class did not enforce proper validation for keyword inputs.
This could result in empty inputs being keyed in.

Let's introduce validation methods for keyword input to ensure that keyword is not empty or null

By enforcing validation, we can improve the robustness of the system and prevent
empty keywrd search